dm-verity: use device write lock around activation
If the same device is attempted to be activated at the exact same
time from concurrent callers, various types of errors will be
returned, depending on the libdevmapper version. This forces all
users to implement complex, costly and error-prone retry logic.

This has been a source of many issues in systemd and libmount,
where the dm-verity device sharing logic is used to avoid
multiple activations, and we really want a definitive and durable
solution.

Use a device write lock around the verity activation, so that these
race conditions cannot happen, and the callers can greatly simplify
their logic.