Changes to support PHMAC with integritysetup
Make the PHMAC integrity algorithm know to libcryptsetup.

The size of a key for PHMAC is not known, because PHMAC gets an opaque
blob as key, who's physical size has nothing to do with the cryptographic
size. Thus, let INTEGRITY_key_size() return -1 as key size for PHMAC, to
indicate that the size is unknown.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>