Add the integrity key size to LUKS2 header if wrapped key HMAC is used
The integrity element of the segment can optionally contain an integer
value called 'key_size' that specifies the size of the integrity key.

In case a wrapped key HMAC algorithm is used, the integrity key size is
passed to LUKS2_generate_hdr() which adds it to the LUSK2 header, in
case it is non-zero.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>