libdevmapper: Detect if dm-crypt supports integrity_key_size option
Check for dm-crypt version 1.28.0 or later. Since this version, dm-crypt
support the integrity_key_size option to pass the size of the integrity key
to the dm-crypt layer.

For LUKS2 with integrity option using a wrapped key HMAC algorithm, the
integrity key is an opaque key blob who's size has nothing to do with the
cryptographic size. dm-rypt would normally use the digest size of the
HMAC algorithm as key size. For wrapped key HMAC this is wrong. Pass the
size of the integrity key to dm-crypt using the integrity_key_size option
instead.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>