cryptsetup: enable use of --volume-key-keyring during 'luksFormat'
Enable the user of --volume-key-keyring with a kernel keystring (e.g.
':32:trusted:data-crypt') when calling luksFormat, which then has
cryptsetup and lib/setup *not* generate a random volume-key, but
instead make use of the linux kernels trusted keyring [1]

The opaque 'trusted' key is handled by passing around the input
 <key_string> inside the volume_key->key_description, and the
 volume_key->key holds the combination of <key_string>::<keyblob>
to be stored in the luks-headers key-slot.

A usage example would look like this:
$> cryptsetup luksFormat --type luks2 /dev/mmcblk2p7 \
  --volume-key-keyring ':32:trusted:data-crypt' --sector-size 4096 \
  --integrity aead --cipher aes-gcm-random:sha256 \
  --key-file /sys/devices/soc0/serial_number --debug --batch-mode

And the corresponding 'open':
$> cryptsetup open /dev/mmcblk2p7 data \
  --key-file /sys/devices/soc0/serial_number --debug

Sidenote: making use of a world-readable 'key-file' is safe, since
the 'trusted' keys are bound to the hardware and are never exposed.

Link 1: https://docs.kernel.org/security/keys/trusted-encrypted.html
Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>